fix: rounding problem

+++ src/ray_casting/horizontal_wall_distance.c
@@ -6,22 +6,22 @@
 /*   By: alida-si <alida-si@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/29 11:41:15 by alida-si          #+#    #+#             */
-/*   Updated: 2023/03/30 13:10:02 by alida-si         ###   ########.fr       */
+/*   Updated: 2023/03/30 13:31:10 by alida-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-int	distance_to_horizontal_wall_up(t_game *game, float angle)
+float	distance_to_horizontal_wall_up(t_game *game, float angle)
 {
-	int	y_coordinate;
-	int	x_coordinate;
-	int	x_step;
-	int	y_step;
-	int	ray_length;
+	float	y_coordinate;
+	float	x_coordinate;
+	float	x_step;
+	float	y_step;
+	float	ray_length;
 
 	y_coordinate = ((int)game->player->y_position / 40) * 40 - 1;
-	x_coordinate = ((int)(game->player->y_position - y_coordinate)
+	x_coordinate = (((int)(game->player->y_position - y_coordinate))
 			/ (-tan(angle))) + game->player->x_position;
 	x_step = 40 / -tan(angle);
 	y_step = -40;
@@ -29,16 +29,16 @@ int	distance_to_horizontal_wall_up(t_game *game, float angle)
 	return (ray_length);
 }
 
-int	distance_to_horizontal_wall_down(t_game *game, float angle)
+float	distance_to_horizontal_wall_down(t_game *game, float angle)
 {
-	int	y_coordinate;
-	int	x_coordinate;
-	int	x_step;
-	int	y_step;
-	int	ray_length;
+	float	y_coordinate;
+	float	x_coordinate;
+	float	x_step;
+	float	y_step;
+	float	ray_length;
 
 	y_coordinate = ((int)game->player->y_position / 40) * 40 + 40;
-	x_coordinate = ((int)(game->player->y_position - y_coordinate)
+	x_coordinate = (((int)(game->player->y_position - y_coordinate))
 			/ (-tan(angle))) + game->player->x_position;
 	x_step = 40 / tan(angle);
 	y_step = 40;
