+++ src/ray_casting/vertical_wall_distance.c
@@ -6,22 +6,22 @@
 /*   By: alida-si <alida-si@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/29 11:39:01 by alida-si          #+#    #+#             */
-/*   Updated: 2023/03/30 09:39:47 by alida-si         ###   ########.fr       */
+/*   Updated: 2023/03/30 13:30:56 by alida-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-int	distance_to_vertical_wall_right(t_game *game, float angle)
+float	distance_to_vertical_wall_right(t_game *game, float angle)
 {
-	int	y_coordinate;
-	int	x_coordinate;
-	int	x_step;
-	int	y_step;
-	int	ray_length;
+	float	y_coordinate;
+	float	x_coordinate;
+	float	x_step;
+	float	y_step;
+	float	ray_length;
 
 	x_coordinate = ((int)game->player->x_position / 40) * 40 + 40;
-	y_coordinate = (int)(game->player->x_position - x_coordinate)
+	y_coordinate = ((int)(game->player->x_position - x_coordinate))
 		* (-tan (angle)) + game->player->y_position;
 	x_step = 40;
 	y_step = 40 * tan(angle);
@@ -29,16 +29,16 @@ int	distance_to_vertical_wall_right(t_game *game, float angle)
 	return (ray_length);
 }
 
-int	distance_to_vertical_wall_left(t_game *game, float angle)
+float	distance_to_vertical_wall_left(t_game *game, float angle)
 {
-	int	y_coordinate;
-	int	x_coordinate;
-	int	x_step;
-	int	y_step;
-	int	ray_length;
+	float	y_coordinate;
+	float	x_coordinate;
+	float	x_step;
+	float	y_step;
+	float	ray_length;
 
 	x_coordinate = ((int)game->player->x_position / 40) * 40 - 1;
-	y_coordinate = (int)(game->player->x_position - x_coordinate)
+	y_coordinate = ((int)(game->player->x_position - x_coordinate))
 		* (-tan (angle)) + game->player->y_position;
 	x_step = -40;
 	y_step = 40 * -tan(angle);
